update views count logic

+++ main-service/src/main/java/com/practice/mainsvc/dto/event/EventFullDto.java
@@ -39,7 +39,7 @@ public class EventFullDto {
 
     String title;
 
-    int views;
+    long views;
 
 
     public void setLocationLat(Double latValue) {
